// src/schemas/player.schema.d.ts
export declare const CreatePlayerSchema: z.ZodObject<{
    name: z.ZodString;
    slug: z.ZodString;
    sector: z.ZodEnum<["BANKING", "TELECOMMUNICATIONS", "ENERGY", "MANUFACTURING", "AGRICULTURE", "REAL_ESTATE", "GOVERNMENT", "REGULATION", "DIVERSIFIED", "INSURANCE", "FINTECH", "RETAIL", "MEDIA", "TRANSPORT"]>;
    type: z.ZodEnum<["LISTED_COMPANY", "SOE", "REGULATOR", "MINISTRY", "PRIVATE_COMPANY", "SUBSIDIARY", "INTERNATIONAL_ORG", "SACCO", "BANK"]>;
    subtype: z.ZodString;
    founded: z.ZodOptional<z.ZodNumber>;
    hq: z.ZodOptional<z.ZodString>;
    ownership: z.ZodOptional<z.ZodString>;
    revenue: z.ZodOptional<z.ZodString>;
    employees: z.ZodOptional<z.ZodString>;
    marketCap: z.ZodOptional<z.ZodString>;
    description: z.ZodString;
    keyFacts: z.ZodDefault<z.ZodOptional<z.ZodArray<z.ZodString, "many">>>;
    tags: z.ZodDefault<z.ZodOptional<z.ZodArray<z.ZodString, "many">>>;
    riskLevel: z.ZodDefault<z.ZodOptional<z.ZodEnum<["LOW", "MEDIUM", "HIGH", "CRITICAL"]>>>;
}, "strip", z.ZodTypeAny, {
    type: "LISTED_COMPANY" | "SOE" | "REGULATOR" | "MINISTRY" | "PRIVATE_COMPANY" | "SUBSIDIARY" | "INTERNATIONAL_ORG" | "SACCO" | "BANK";
    sector: "BANKING" | "TELECOMMUNICATIONS" | "ENERGY" | "MANUFACTURING" | "AGRICULTURE" | "REAL_ESTATE" | "GOVERNMENT" | "REGULATION" | "DIVERSIFIED" | "INSURANCE" | "FINTECH" | "RETAIL" | "MEDIA" | "TRANSPORT";
    name: string;
    slug: string;
    subtype: string;
    description: string;
    keyFacts: string[];
    tags: string[];
    riskLevel: "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
    founded?: number | undefined;
    hq?: string | undefined;
    ownership?: string | undefined;
    revenue?: string | undefined;
    employees?: string | undefined;
    marketCap?: string | undefined;
}, {
    type: "LISTED_COMPANY" | "SOE" | "REGULATOR" | "MINISTRY" | "PRIVATE_COMPANY" | "SUBSIDIARY" | "INTERNATIONAL_ORG" | "SACCO" | "BANK";
    sector: "BANKING" | "TELECOMMUNICATIONS" | "ENERGY" | "MANUFACTURING" | "AGRICULTURE" | "REAL_ESTATE" | "GOVERNMENT" | "REGULATION" | "DIVERSIFIED" | "INSURANCE" | "FINTECH" | "RETAIL" | "MEDIA" | "TRANSPORT";
    name: string;
    slug: string;
    subtype: string;
    description: string;
    founded?: number | undefined;
    hq?: string | undefined;
    ownership?: string | undefined;
    revenue?: string | undefined;
    employees?: string | undefined;
    marketCap?: string | undefined;
    keyFacts?: string[] | undefined;
    tags?: string[] | undefined;
    riskLevel?: "LOW" | "MEDIUM" | "HIGH" | "CRITICAL" | undefined;
}>;
// ...
export declare const GetPlayersSchema: z.ZodObject<{
    sector: z.ZodOptional<z.ZodEnum<["BANKING", "TELECOMMUNICATIONS", "ENERGY", "MANUFACTURING", "AGRICULTURE", "REAL_ESTATE", "GOVERNMENT", "REGULATION", "DIVERSIFIED", "INSURANCE", "FINTECH", "RETAIL", "MEDIA", "TRANSPORT"]>>;
    type: z.ZodOptional<z.ZodEnum<["LISTED_COMPANY", "SOE", "REGULATOR", "MINISTRY", "PRIVATE_COMPANY", "SUBSIDIARY", "INTERNATIONAL_ORG", "SACCO", "BANK"]>>;
    tags: z.ZodOptional<z.ZodString>;
    search: z.ZodOptional<z.ZodString>;
    riskLevel: z.ZodOptional<z.ZodEnum<["LOW", "MEDIUM", "HIGH", "CRITICAL"]>>;
    page: z.ZodDefault<z.ZodNumber>;
    limit: z.ZodDefault<z.ZodNumber>;
}, "strip", z.ZodTypeAny, {
    page: number;
    limit: number;
    type?: "LISTED_COMPANY" | "SOE" | "REGULATOR" | "MINISTRY" | "PRIVATE_COMPANY" | "SUBSIDIARY" | "INTERNATIONAL_ORG" | "SACCO" | "BANK" | undefined;
    sector?: "BANKING" | "TELECOMMUNICATIONS" | "ENERGY" | "MANUFACTURING" | "AGRICULTURE" | "REAL_ESTATE" | "GOVERNMENT" | "REGULATION" | "DIVERSIFIED" | "INSURANCE" | "FINTECH" | "RETAIL" | "MEDIA" | "TRANSPORT" | undefined;
    search?: string | undefined;
    tags?: string | undefined;
    riskLevel?: "LOW" | "MEDIUM" | "HIGH" | "CRITICAL" | undefined;
}, {
    type?: "LISTED_COMPANY" | "SOE" | "REGULATOR" | "MINISTRY" | "PRIVATE_COMPANY" | "SUBSIDIARY" | "INTERNATIONAL_ORG" | "SACCO" | "BANK" | undefined;
    sector?: "BANKING" | "TELECOMMUNICATIONS" | "ENERGY" | "MANUFACTURING" | "AGRICULTURE" | "REAL_ESTATE" | "GOVERNMENT" | "REGULATION" | "DIVERSIFIED" | "INSURANCE" | "FINTECH" | "RETAIL" | "MEDIA" | "TRANSPORT" | undefined;
    search?: string | undefined;
    tags?: string | undefined;
    riskLevel?: "LOW" | "MEDIUM" | "HIGH" | "CRITICAL" | undefined;
    page?: number | undefined;
    limit?: number | undefined;
}>;